refactor: graduate ConnectPhase to @kolu/surface-remote (framework), adopt in client
srid's upgrade to the type-dedup: ConnectPhase (the up-but-not-yet-connected phase
subset a connect/progress UI narrates) now lives in the FRAMEWORK, beside its honest
owner ConnectionInfo, instead of being derived in the kolu client. A surface-API
addition, so a drishti pair PR adopts it (second-consumer demand proof).

- @kolu/surface-remote/connection: export ConnectPhase =
  Exclude<ConnectionInfo['phase'], 'connected'|'disconnected'|'failed'> (auto-tracks
  SshProv). Exported from the package index + re-exported through
  kolu-common/surfacesWithPadi (the client's established path). Surface reference
  updated.
- client connectCanvasCopy.ts: import ConnectPhase — the local hand-list is deleted.
- canvasModeResolver: the not-yet-connected arms type connectPhase as ConnectPhase
  (TIGHTER than the full union — connected/disconnected/failed unconstructible); the
  narrowing moves to useCanvasMode (isConnectPhase at the facts boundary), so bindingUp
  collapses to '!== undefined'. @ts-expect-error pins now prove both 'banana' AND
  'connected' are unconstructible on the arm.

Left untouched per the review: EntryConnectionState coarsening, HostReconcileAction,
createRejoinKeyedSub, floorConnectionInfo.
