+++ packages/client/src/kaval/connectCanvasCopy.ts
@@ -10,19 +10,11 @@
  *  `disconnected`/`failed` are owned by the Skew-UX host-down card — NOT narrated here
  *  (a second failure surface is exactly what this must not build). */
 
-import type { ConnectionInfo } from "kolu-common/surfacesWithPadi";
-
-/** The phases the connect overlay narrates — the UP-but-not-yet-connected subset of
- *  `ConnectionInfo["phase"]`, DERIVED from the one source of truth rather than hand-listed:
- *  it resolves to exactly `SshProv | "connecting"` today and auto-tracks any future `SshProv`
- *  provisioning phase. Add a phase to `SshProv` and {@link connectCanvasCopy}'s switch fails
- *  to compile (missing case) — the drift signal a hand-copied literal union silently swallowed
- *  (the connect overlay would just stop narrating the new phase). Type-only import, so this
- *  stays browser-safe and adds no runtime dependency. */
-export type ConnectPhase = Exclude<
-  ConnectionInfo["phase"],
-  "connected" | "disconnected" | "failed"
->;
+// The phases the connect overlay narrates ride the framework's own `ConnectPhase` (exported
+// beside `ConnectionInfo`, its honest owner) — imported through kolu-common's established
+// re-export, NOT re-listed here. Adding an `SshProv` provisioning phase then fails
+// {@link connectCanvasCopy}'s switch to compile (missing case) — the drift signal.
+import type { ConnectPhase } from "kolu-common/surfacesWithPadi";
 
 export interface ConnectCopy {
   /** The headline line. */
