+++ website/src/content/surface/ref-surface-remote.mdx
@@ -100,6 +100,7 @@ up, so `failed` + `network` is unrepresentable). There is no separate
 | --- | --- | --- |
 | `mirroredSurface(base)` | `/connection` | the public composition path: adds the gate-closed, get-only `connection` cell and reserves the name (throws on a base that already declares one) |
 | `connectionCell` / `ConnectionInfoSchema` / `DEFAULT_CONNECTION` | `/connection` | the exported fragment — read to seed a store or assert the shape, not to spread yourself |
+| `ConnectPhase` | `/connection` | the up-but-not-yet-connected phase subset a connect/progress UI narrates — `Exclude<ConnectionInfo["phase"], "connected" \| "disconnected" \| "failed">` (i.e. `SshProv \| "connecting"`), the ONE vocabulary a UI's exhaustive `switch`/`Record` keys on so a new provisioning phase fails to compile until handled (never a silent hand-listed copy) |
 | `LogEntry` / `LogEntrySchema` | `/connection` | one `{ source, line }` log-tail entry — the browser type of a `SessionState.log` line |
 | `pipeSessionStateToCell(session, set)` | `/connection` | the node-side pump that drives the cell off `session.onState` (`pumpRemoteSurface` wires it for you when the mirrored surface carries `connection`) |
